Extract CastParameters module
